@@ -1,6 +1,6 @@
 [metadata]
 name = findssh
-version = 1.4.1
+version = 1.5.0
 author = Michael Hirsch, Ph.D.
 author_email = scivision@users.noreply.github.com
 url = https://github.com/scivision/findssh
@@ -19,11 +19,7 @@ classifiers =
   Intended Audience :: System Administrators
   Intended Audience :: Telecommunications Industry
   Operating System :: OS Independent
-  Programming Language :: Python :: 3.5
-  Programming Language :: Python :: 3.6
-  Programming Language :: Python :: 3.7
-  Programming Language :: Python :: 3.8
-  Programming Language :: Python :: 3.9
+  Programming Language :: Python :: 3
   Topic :: Security
   Topic :: System :: Networking
   Topic :: Utilities
@@ -33,7 +29,7 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 
 [options]
-python_requires = >= 3.5
+python_requires = >= 3.7
 packages = find:
 zip_safe = False
 include_package_data = True
