@@ -1,4 +1,5 @@
 import ipaddress
+import asyncio
 
 import findssh
 
@@ -9,7 +10,7 @@ TIMEOUT = 1.0
 
 def test_coroutine():
     net = findssh.netfromaddress(findssh.getLANip())
-    hosts = findssh.runner(findssh.get_hosts, net, PORT, SERVICE, TIMEOUT)
+    hosts = asyncio.run(findssh.get_hosts(net, PORT, SERVICE, TIMEOUT))
     if len(hosts) > 0:
         host = hosts[0]
         assert isinstance(host[0], ipaddress.IPv4Address)
